# docs/swagger.yaml
basePath: /api
definitions:
  api.BusinessError:
    properties:
      code:
        description: Code error code, 0 indicates success, otherwise business error
          (e.g. user not found) or internal server error.
        type: integer
      data:
        description: Data is the return value if success. Otherwise, it is concret
          failure reason in string type.
      message:
        description: Message error message associated with `Code`.
        type: string
    type: object
  api.GasTankConfig:
    properties:
      paymaster:
        description: Paymaster is the address of the gas tank paymaster contract.
        items:
          type: integer
        type: array
      tokens:
        items:
          type: string
        type: array
    type: object
  api.GasTankStubRequest:
    properties:
      sender:
        description: Smart account address in hex format with 0x prefix.
        type: string
      token:
        description: ERC20 token address to pay gas fee.
        type: string
    required:
    - sender
    - token
    type: object
  api.PaymasterAndDataStub:
    properties:
      address:
        type: string
      data:
        type: string
    type: object
  api.SetCodeAuth:
    properties:
      chainId:
        description: ChainID is the chain ID the authorization is bound to. 0 means
          the auth is valid on any chain.
        type: integer
      contract:
        description: |-
          Contract is the 20-byte hex address (with 0x prefix) of the smart-contract to delegate the EOA to.
          Set to "0x0000000000000000000000000000000000000000" to revoke an existing delegation.
        type: string
      nonce:
        description: Nonce is the current on-chain nonce of the signing authority
          (EOA). Must match exactly.
        type: integer
      r:
        description: R is the R component of the EIP-7702 authorization signature,
          as a 0x-prefixed hex string.
        type: string
      s:
        description: S is the S component of the EIP-7702 authorization signature,
          as a 0x-prefixed hex string.
        type: string
      v:
        description: V is the recovery identifier of the EIP-7702 authorization signature
          (0 or 1).
        type: integer
    required:
    - contract
    - r
    - s
    type: object
  api.SetCodeResult:
    properties:
      error:
        description: Error contains the failure reason reported by the EVM when Success
          is false.
        type: string
      executed:
        description: Executed indicates whether the transaction has been executed.
        type: boolean
      success:
        description: Success indicates whether the set-code authorization succeeded.
          Only meaningful when Executed is true.
        type: boolean
    type: object
  api.TokenPayConfig:
    properties:
      maxGasCost:
        description: MaxGasCost is the maximum allowed gas cost in wei.
        type: integer
      minGasFeeRatio:
        description: MinGasFeeRatio is the minimum allowed gas fee ratio (gas fee
          / base fee) in percentage. For example, 120 means the gas fee must be at
          least 120% of the base fee.
        type: integer
      minGasTipRatio:
        description: MinGasTipRatio is the minimum allowed gas tip ratio (priority
          fee / base fee) in percentage. For example, 20 means the priority fee must
          be at least 20% of the base fee.
        type: integer
      recipient:
        description: Recipient is the configured recipient address that receives token
          payments.
        type: string
      suggestedGasPriceBumpRatio:
        description: SuggestedGasPriceBumpRatio is the percentage by which to bump
          the suggested gas price to increase the chance of timely inclusion in blocks.
        type: integer
      suggestedTokenPriceBumpRatio:
        description: SuggestedTokenPriceBumpRatio is the percentage by which to bump
          the suggested token price to increase the chance of timely inclusion in
          blocks when the price is volatile.
        type: integer
      tokens:
        description: Tokens is the list of ERC20 token contracts supported for token-pay.
          Note, the tokens[0] is the default USDT token used for quoting and payment.
        items:
          type: string
        type: array
    type: object
  api.TokenPayRequest:
    properties:
      rawBusinessTx:
        type: string
      rawTransferTokenTx:
        type: string
    required:
    - rawBusinessTx
    - rawTransferTokenTx
    type: object
  api.UserOperation:
    properties:
      callData:
        minLength: 2
        type: string
      callGasLimit:
        maxLength: 32
        type: string
      factory:
        type: string
      factoryData:
        minLength: 2
        type: string
      maxFeePerGas:
        maxLength: 32
        type: string
      maxPriorityFeePerGas:
        maxLength: 32
        type: string
      nonce:
        type: string
      paymaster:
        description: Paymaster
        type: string
      paymasterData:
        description: at least validAfter (6) || validUntil (6) || signature (65)
        minLength: 156
        type: string
      paymasterPostOpGasLimit:
        maxLength: 32
        type: string
      paymasterVerificationGasLimit:
        maxLength: 32
        type: string
      preVerificationGas:
        maxLength: 32
        type: string
      sender:
        type: string
      signature:
        type: string
      verificationGasLimit:
        maxLength: 32
        type: string
    required:
    - callData
    - callGasLimit
    - maxFeePerGas
    - maxPriorityFeePerGas
    - nonce
    - paymaster
    - paymasterData
    - paymasterPostOpGasLimit
    - paymasterVerificationGasLimit
    - preVerificationGas
    - sender
    - signature
    - verificationGasLimit
    type: object
  api.VerifyingPaymasterConfig:
    properties:
      contractWhitelist:
        items:
          type: string
        type: array
      maxGasCost:
        type: integer
      smartAccountWhitelist:
        items:
          type: string
        type: array
    type: object
  api.VerifyingPaymasterStubRequest:
    properties:
      delegation:
        description: Delegated contract address in hex format with 0x prefix. Set
          to non-zero value if user op carries a 7702 auth message.
        type: string
      sender:
        description: Smart account address in hex format with 0x prefix.
        type: string
    required:
    - delegation
    - sender
    type: object
info:
  contact: {}
  description: Fluent Backend provides account abstraction, gas tank paymaster, and
    ERC20 token gas payment services.
  title: Fluent Backend API
  version: "1.0"
paths:
  /aa/auth:
    post:
      consumes:
      - application/json
      description: |-
        Validates the signed EIP-7702 authorization and broadcasts a set-code transaction.
        The service will sponsor the set-code transaction gas fee, and user needs to poll GET /aa/auth/{txHash} for the result.
      operationId: aaAuthSend
      parameters:
      - description: Signed EIP-7702 authorization message
        in: body
        name: auth
        required: true
        schema:
          $ref: '#/definitions/api.SetCodeAuth'
      produces:
      - application/json
      responses:
        "200":
          description: Transaction hash (0x-prefixed hex)
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                data:
                  type: string
              type: object
        "600":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                data:
                  type: string
              type: object
      summary: Submit EIP-7702 authorization
      tags:
      - AccountAbstract
  /aa/auth/{txHash}:
    get:
      consumes:
      - application/json
      description: |-
        Returns the executed result of an EIP-7702 set-code transaction. Returns executed=false while the transaction is still pending.
        If the transaction is not found (business error code 1001), re-submit via POST /aa/auth.
      operationId: aaAuthStatus
      parameters:
      - description: 0x-prefixed 32-byte transaction hash (66 characters)
        in: path
        name: txHash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Set-code result (check executed and success fields)
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                data:
                  $ref: '#/definitions/api.SetCodeResult'
              type: object
        "600":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                data:
                  type: string
              type: object
      summary: Query EIP-7702 set-code result
      tags:
      - AccountAbstract
  /aa/gastank/config:
    get:
      consumes:
      - application/json
      description: Returns the configuration of the gas tank paymaster.
      operationId: aaGasTankConfig
      produces:
      - application/json
      responses:
        "200":
          description: Gas tank paymaster configuration
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                data:
                  $ref: '#/definitions/api.GasTankConfig'
              type: object
        "600":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                data:
                  type: string
              type: object
      summary: Get gas tank paymaster configuration
      tags:
      - GasTank
  /aa/gastank/sign:
    post:
      consumes:
      - application/json
      description: |-
        Calculates maxTokenCost for the given UserOperation, adds paymaster signature, and returns reassembled paymasterData.
        Encoding format (129 bytes): token(20) || maxTokenCost(32) || validAfter(6) || validUntil(6) || signature(65). maxTokenCost is hex encoded.
      operationId: aaGasTankSign
      parameters:
      - description: UserOperation for maxTokenCost estimation and paymaster signing
        in: body
        name: userOp
        required: true
        schema:
          $ref: '#/definitions/api.UserOperation'
      produces:
      - application/json
      responses:
        "200":
          description: Signed and reassembled paymasterData
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                data:
                  type: string
              type: object
        "600":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                data:
                  type: string
              type: object
      summary: Sign paymasterData for UserOperation
      tags:
      - GasTank
  /aa/gastank/stub:
    post:
      consumes:
      - application/json
      description: Returns paymasterData used for user operation gas estimation.
      operationId: aaGasTankStub
      parameters:
      - description: Paymaster data stub request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.GasTankStubRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Paymaster address and data (0x-prefixed hex)
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                data:
                  $ref: '#/definitions/api.PaymasterAndDataStub'
              type: object
        "600":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                data:
                  type: string
              type: object
      summary: Prepare paymasterData for gas estimation
      tags:
      - GasTank
  /aa/paymaster/config:
    get:
      consumes:
      - application/json
      description: Returns the configuration of the verifying paymaster.
      operationId: aaPaymasterConfig
      produces:
      - application/json
      responses:
        "200":
          description: Verifying paymaster configuration
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                data:
                  $ref: '#/definitions/api.VerifyingPaymasterConfig'
              type: object
        "600":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                data:
                  type: string
              type: object
      summary: Get verifying paymaster configuration
      tags:
      - Paymaster
  /aa/paymaster/sign:
    post:
      consumes:
      - application/json
      description: Validates the given UserOperation, adds paymaster signature, and
        returns reassembled paymasterData.
      operationId: aaPaymasterSign
      parameters:
      - description: UserOperation for paymaster signing
        in: body
        name: userOp
        required: true
        schema:
          $ref: '#/definitions/api.UserOperation'
      produces:
      - application/json
      responses:
        "200":
          description: Signed and reassembled paymasterData (0x-prefixed hex)
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                data:
                  type: string
              type: object
        "600":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                data:
                  type: string
              type: object
      summary: Sign paymasterData of given user operation and return reassembled paymasterData
      tags:
      - Paymaster
  /aa/paymaster/stub:
    post:
      consumes:
      - application/json
      description: Returns the stub paymasterData of verifying paymaster for gas estimation.
      operationId: aaPaymasterStub
      parameters:
      - description: Verifying paymaster stub request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.VerifyingPaymasterStubRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Paymaster address and data (0x-prefixed hex)
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                data:
                  $ref: '#/definitions/api.PaymasterAndDataStub'
              type: object
        "600":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                data:
                  type: string
              type: object
      summary: Returns the stub paymasterData of verifying paymaster for gas estimation
      tags:
      - Paymaster
  /tokenpay/config:
    get:
      consumes:
      - application/json
      description: Returns the supported ERC20 token list and the configured recipient
        address for token-pay.
      operationId: tokenPayConfig
      produces:
      - application/json
      responses:
        "200":
          description: Token-pay configuration
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                data:
                  $ref: '#/definitions/api.TokenPayConfig'
              type: object
        "600":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                data:
                  type: string
              type: object
      summary: Get token-pay configuration
      tags:
      - TokenPay
  /tokenpay/price:
    get:
      consumes:
      - application/json
      description: Returns the ETH price quoted in the specified token, i.e. how many
        token units are required for 1 ETH.
      operationId: tokenPayGetETHPrice
      parameters:
      - description: ERC20 token address (0x-prefixed, 42 characters)
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ETH price in token smallest units
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                data:
                  type: string
              type: object
        "600":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                data:
                  type: string
              type: object
      summary: Get ETH price by token
      tags:
      - TokenPay
  /tokenpay/submit:
    post:
      consumes:
      - application/json
      description: |-
        Submits two raw transactions (transfer token and business) for sponsored execution.
        Backend validates both transactions first. If validation passes, it sends a funding ETH transaction.
        After funding is confirmed asynchronously, backend broadcasts transfer-token tx then business tx.
        Clients can poll on-chain status of both submitted transactions (transfer-token and business) by their tx hashes.
        The whole flow is considered completed only when both transactions are successfully executed.
        If transfer-token execution fails, the user may be blacklisted by risk control.
        If transactions remain pending for a long time, contact administrator for investigation.
      operationId: tokenPaySubmit
      parameters:
      - description: Raw transfer-token tx and raw business tx
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.TokenPayRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Submission accepted
          schema:
            $ref: '#/definitions/api.BusinessError'
        "600":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                data:
                  type: string
              type: object
      summary: Submit token-pay transactions
      tags:
      - TokenPay
swagger: "2.0"
tags:
- description: Provides account abstraction features, including free EOA to smart
    account upgrades and upgrade status queries.
  name: AccountAbstract
- description: Provides paymaster services for gas fee sponsorship.
  name: Paymaster
- description: Provides an off-chain gas tank sponsorship service for smart accounts,
    including paymaster data preparation and user operation signing for on-chain paymaster
    contract validation.
  name: GasTank
- description: Provides gas fee sponsorship so EOA accounts can pay gas fees with
    supported ERC20 tokens.
  name: TokenPay